feat: Add sign-in with manager functionality
This commit adds the functionality to sign in with a manager account. The `signInWithManager` method is implemented in the `LoginController` class. It retrieves the user credentials using the `signInWithEmailAndPassword` method from Firebase Authentication. If the user credentials are valid, the `setUserByNow` method is called to set the user. Finally, if the context is still mounted, the user is redirected to the main page.